% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiple_uncertainty.R
\name{multiple_uncertainty}
\alias{multiple_uncertainty}
\title{multiple uncertainty}
\usage{
multiple_uncertainty(f = logistic, x_grid = seq(0, 150, length = 151),
  h_grid = x_grid, Tmax = 50, sigma_g = 0.2, sigma_m = 0, sigma_i = 0,
  delta = 0.05, noise_dist = "uniform", y_grid = x_grid,
  q_grid = x_grid)
}
\description{
Stochastic dynamic programming solution under multiple uncertainty
}

